requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
accelerate>=0.21.0
unsloth>=0.3.0
einops>=0.6.0
open-clip-torch>=2.20.0
timm>=1.0.0

# Data processing
datasets>=2.10.0
evaluate>=0.4.0
huggingface-hub>=0.19.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
ftfy>=6.1.0

# Visualization and UI
streamlit>=1.25.0
plotly>=5.15.0
matplotlib>=3.7.0

# Experiment tracking
wandb>=0.15.0

# Optimization tools
bitsandbytes>=0.40.0
optimum>=1.12.0
onnxruntime>=1.15.0

# EfficientVLM
git+https://github.com/swaggy-TN/EfficientVLM.git

# TinyChat for 4-bit quantization
git+https://github.com/mit-han-lab/tinychat.git